@@ -4,10 +4,10 @@ import Button from './Button';
 import Notification from './Notification';
 import Portal from './Portal';
 import { User, Notebook, Map } from './icons';
-import mapPin from 'assets/images/map_pin.svg';
-import mapPinShine from 'assets/images/map_pin_shine.svg';
-import unfocusedPin from 'assets/images/unfocused_pin.svg';
-import unfocusedPinPulse from 'assets/images/unfocused_pin_pulse.svg';
+import mapPin from '@src/assets/images/map_pin.svg';
+import mapPinShine from '@src/assets/images/map_pin_shine.svg';
+import unfocusedPin from '@src/assets/images/unfocused_pin.svg';
+import unfocusedPinPulse from '@src/assets/images/unfocused_pin_pulse.svg';
 import './DebugZone.scss';
 
 /**
