@@ -1,26 +1,25 @@
 import './Navigation.scss';
-import React from 'react';
 import { NavLink } from 'react-router-dom';
 import { User, Notebook, Map } from './icons';
-import Routes from '../types/Routes';
+import Routes from '@src/types/Routes';
 
 const Navigation = () => {
   return (
     <nav className="Navigation">
       <ul>
-        <NavLink activeClassName="active" to={Routes.Profile}>
+        <NavLink to={Routes.Profile}>
           <li>
             <User />
             <span>Profile</span>
           </li>
         </NavLink>
-        <NavLink activeClassName="active" to={Routes.Map}>
+        <NavLink to={Routes.Map}>
           <li>
             <Map />
             <span>Map</span>
           </li>
         </NavLink>
-        <NavLink activeClassName="active" to={Routes.Friends}>
+        <NavLink to={Routes.Friends}>
           <li>
             <Notebook />
             <span>Friends</span>
