@@ -1,6 +1,5 @@
 import React, { useEffect, useState, useContext } from 'react';
-import firebaseContext from 'firebaseContext';
-import API from 'api/axios';
+import API from '@src/api/axios';
 import InputText from './InputText';
 import Button from './Button';
 import * as authApi from '../api/auth';
@@ -10,7 +9,6 @@ const SignUp: React.FC = () => {
   const [error, setError] = useState('');
   const [email, updateEmail] = useState('');
   const [password, updatePassword] = useState('');
-  const firebase = useContext(firebaseContext);
 
   useEffect(() => {
     if (email && password) {
@@ -25,18 +23,7 @@ const SignUp: React.FC = () => {
 
   async function handleSignUp(e: React.FormEvent<HTMLFormElement>) {
     e.preventDefault();
-    try {
-      await authApi.signUp(email, password);
-      await firebase.auth().signInWithEmailAndPassword(email, password);
-      const token = await firebase.auth().currentUser?.getIdToken();
-      API.defaults.headers['Authorization'] = `Bearer ${token}`;
-    } catch (e) {
-      if (e.message === 'Request failed with status code 404') {
-        handleError(new Error('Email is already registered'));
-      } else {
-        handleError(e);
-      }
-    }
+    await authApi.signUp(email, password);
   }
 
   return (
