@@ -1,4 +1,4 @@
-import { ToolColor } from 'types';
+import { ToolColor } from '@src/types';
 
 export function backgroundColor(color: ToolColor): string {
   return `background-${color}`;
